Fit a trending model

fit will return the output from fitting the model to the specified data.

# S3 method for trending_model
fit(object, data, ...)

Arguments

object
A trending_model object (i.e. the output from function lm_model, glm_model, glm.nb_model, or brm_model).

data
A data frame containing the data to fit.

...
Not currently used.

Value

A trending_fit object which is a list subclass with entries:

result: the resulting fit from calling the underlying model directly, i.e.

lm_model: a fitted model object of class lm

glm_model: a fitted model object of class glm

glm.nb_model: a fitted model object of class negbin

brm_model: An object of class brmsfit

NULL if fitting fails.

warnings: any warnings generated during fitting

errors: any errors generated during fitting

Author

Tim Taylor

Examples

x = rnorm(100, mean = 0)
y = rpois(n = 100, lambda = exp(1.5 + 0.5*x))
dat <- data.frame(x = x, y = y)

poisson_model <- glm_model(y ~ x , family = poisson)
negbin_model <- glm.nb_model(y ~ x)

fit(poisson_model, dat)

#> $result
#> 
#> Call:  glm(formula = y ~ x, family = poisson, data = dat)
#> 
#> Coefficients:
#> (Intercept)            x  
#>      1.5024       0.5486  
#> 
#> Degrees of Freedom: 99 Total (i.e. Null);  98 Residual
#> Null Deviance:	    265.4 
#> Residual Deviance: 98.39 	AIC: 421.9
#> 
#> $warnings
#> NULL
#> 
#> $errors
#> NULL
#> 
#> attr(,"class")
#> [1] "trending_fit" "list"        

fit(negbin_model, dat)

#> $result
#> 
#> Call:  glm.nb(formula = y ~ x, data = dat, init.theta = 55556.32135, 
#>     link = log)
#> 
#> Coefficients:
#> (Intercept)            x  
#>      1.5024       0.5486  
#> 
#> Degrees of Freedom: 99 Total (i.e. Null);  98 Residual
#> Null Deviance:	    265.3 
#> Residual Deviance: 98.38 	AIC: 423.9
#> 
#> $warnings
#> [1] "iteration limit reached" "iteration limit reached"
#> 
#> $errors
#> NULL
#> 
#> attr(,"class")
#> [1] "trending_fit" "list"