Predict method for trending_fit objects

predict() adds estimated values and associated confidence and/or prediction intervals for various fitted models.

# S3 method for trending_fit
predict(
  object,
  new_data,
  name = "estimate",
  alpha = 0.05,
  add_ci = TRUE,
  ci_names = c("lower_ci", "upper_ci"),
  add_pi = TRUE,
  pi_names = c("lower_pi", "upper_pi"),
  simulate_pi = FALSE,
  sims = 2000,
  uncertain = TRUE,
  as_tibble = FALSE,
  ...
)

# S3 method for trending_fit_list
predict(
  object,
  new_data,
  name = "estimate",
  alpha = 0.05,
  add_ci = TRUE,
  ci_names = c("lower_ci", "upper_ci"),
  add_pi = TRUE,
  pi_names = c("lower_pi", "upper_pi"),
  simulate_pi = FALSE,
  sims = 2000,
  uncertain = TRUE,
  as_tibble = FALSE,
  ...
)

# S3 method for trending_fit_tbl
predict(
  object,
  new_data,
  name = "estimate",
  alpha = 0.05,
  add_ci = TRUE,
  ci_names = c("lower_ci", "upper_ci"),
  add_pi = TRUE,
  pi_names = c("lower_pi", "upper_pi"),
  simulate_pi = FALSE,
  sims = 2000,
  uncertain = TRUE,
  as_tibble = FALSE,
  ...
)

Arguments

object
A trending_fit, trending_fit_list or trending_fit_tbl object.

new_data
A data.frame containing data for which estimates are to be derived. If missing, the model frame from the fit data will be used.

name
Character vector of length one giving the name to use for the calculated estimate.

alpha
The alpha threshold to be used for prediction intervals, defaulting to 0.05, i.e. 95% prediction intervals are derived.

add_ci
Should a confidence interval be added to the output. Default TRUE.

ci_names
Names to use for the resulting confidence intervals.

add_pi
Should a prediction interval be added to the output. Default TRUE.

pi_names
Names to use for the resulting prediction intervals.

simulate_pi
Should the prediction intervals for glm models be simulated. If TRUE, default, predict() uses the ciTools::add_pi() function to generate the intervals.

sims
The number of simulations to run when simulating prediction intervals for a glm model.

uncertain
Only used for glm models and when simulate_pi = FALSE. Default TRUE. If FALSE uncertainty in the fitted parameters is ignored when generating the parametric prediction intervals.

as_tibble
Should the output be converted to a tibble subclass.

...
Not currently used.

Value

For trending_fit inputs, if as_tibble = FALSE, a trending_predict object, which is a list subclass, with entries:

result: the input data frame with additional estimates and, optionally, confidence and or prediction intervals. NULL if the associated predict method fails.

warnings: any warnings generated during prediction.

errors: any errors generated during prediction.

If as_tibble = TRUE, a trending_predict_tbl object which is a tibble subclass with one row and columns 'result', 'warnings' and 'errors' with contents as above.

For trending_fit_list or trending_fit_tbl inputs, if as_tibble = FALSE, a trending_fit_list where each entry is a trending_fit object. If as_tibble = TRUE a trending_predict_tbl object with each row corresponding to the output of one model as above.

Examples

x = rnorm(100, mean = 0)
y = rpois(n = 100, lambda = exp(1.5 + 0.5*x))
dat <- data.frame(x = x, y = y)

poisson_model <- glm_model(y ~ x , family = "poisson")
negbin_model <- glm_nb_model(y ~ x)

fitted_poisson <- fit(poisson_model, dat)
fitted_negbin <- fit(negbin_model, dat)
fitted_list <- fit_list(list(poisson_model, negbin_model), dat)
fitted_list_named <- fit_list(list(p = poisson_model, n = negbin_model), dat)

predict(fitted_poisson)

#> $result
#> <trending_estimate> 100 x 7
#>        y       x estimate lower_ci upper_ci lower_pi upper_pi
#>    <int>   <dbl>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#>  1     3 -1.08       2.14     1.80     2.54        0        6
#>  2     1 -1.22       1.96     1.63     2.35        0        6
#>  3     2  0.0849     4.42     4.00     4.88        1       10
#>  4     7  0.581      6.02     5.53     6.55        1       12
#>  5     4  0.451      5.55     5.08     6.05        1       11
#>  6     5  0.497      5.71     5.24     6.23        1       12
#>  7     9  0.779      6.81     6.25     7.41        2       13
#>  8     7  1.51      10.7      9.61    11.9         4       19
#>  9     1 -1.82       1.35     1.07     1.69        0        5
#> 10     3  0.788      6.84     6.29     7.45        2       13
#> # … with 90 more rows
#> 
#> $warnings
#> NULL
#> 
#> $errors
#> NULL
#> 
#> attr(,"class")
#> [1] "trending_predict" "list"            

predict(fitted_list)

#> [[1]]
#> $result
#> NULL
#> 
#> $warnings
#> NULL
#> 
#> $errors
#> [1] "Not implemented for class NULL"
#> 
#> attr(,"class")
#> [1] "trending_predict" "list"            
#> 
#> [[2]]
#> $result
#> NULL
#> 
#> $warnings
#> NULL
#> 
#> $errors
#> [1] "Not implemented for class NULL"
#> 
#> attr(,"class")
#> [1] "trending_predict" "list"            
#> 
#> attr(,"class")
#> [1] "trending_predict_list" "list"                 

predict(fitted_list, as_tibble = TRUE)

#> # A tibble: 2 × 3
#>   result warnings errors   
#>   <list> <list>   <list>   
#> 1 <NULL> <NULL>   <chr [1]>
#> 2 <NULL> <NULL>   <chr [1]>

predict(fitted_list_named)

#> [[1]]
#> $result
#> NULL
#> 
#> $warnings
#> NULL
#> 
#> $errors
#> [1] "Not implemented for class NULL"
#> 
#> attr(,"class")
#> [1] "trending_predict" "list"            
#> 
#> [[2]]
#> $result
#> NULL
#> 
#> $warnings
#> NULL
#> 
#> $errors
#> [1] "Not implemented for class NULL"
#> 
#> attr(,"class")
#> [1] "trending_predict" "list"            
#> 
#> attr(,"class")
#> [1] "trending_predict_list" "list"                 

predict(fitted_list_named, as_tibble = TRUE)

#> # A tibble: 2 × 3
#>   result warnings errors   
#>   <list> <list>   <list>   
#> 1 <NULL> <NULL>   <chr [1]>
#> 2 <NULL> <NULL>   <chr [1]>