Predict method for trending_model objects

Adds estimated values and associated confidence and/or prediction intervals to data based on trending_model fit.

# S3 method for trending_model
predict(
  object,
  data,
  name = "estimate",
  alpha = 0.05,
  add_ci = TRUE,
  ci_names = c("lower_ci", "upper_ci"),
  add_pi = TRUE,
  pi_names = c("lower_pi", "upper_pi"),
  simulate_pi = FALSE,
  sims = 2000,
  uncertain = TRUE,
  as_tibble = TRUE,
  ...
)

Arguments

object
A trending_model object.

data
A data.frame containing data to which the model is to be fit and estimates derived.

name
Character vector of length one giving the name to use for the calculated estimate.

alpha
The alpha threshold to be used for prediction intervals, defaulting to 0.05, i.e. 95% prediction intervals are derived.

add_ci
Should a confidence interval be added to the output. Default TRUE.

ci_names
Names to use for the resulting confidence intervals.

add_pi
Should a prediction interval be added to the output. Default TRUE.

pi_names
Names to use for the resulting prediction intervals.

simulate_pi
Should the prediction intervals for glm models be simulated. If TRUE, default, predict() uses the ciTools::add_pi() function to generate the intervals.

sims
The number of simulations to run when simulating prediction intervals for a glm model.

uncertain
Only used for glm models and when simulate_pi = FALSE. Default TRUE. If FALSE uncertainty in the fitted parameters is ignored when generating the parametric prediction intervals.

as_tibble
Should the output be converted to a tibble subclass.

...
Not currently used.

Value

If as_tibble = FALSE, a trending_predict object, which is a list subclass, with entries:

result: the input data frame with additional estimates and, optionally, confidence and or prediction intervals. NULL if the associated predict method fails.

warnings: any warnings generated during prediction.

errors: any errors generated during prediction.

If as_tibble = TRUE, a trending_predict_tbl object which is a tibble subclass with one row per model and columns 'result', 'warnings' and 'errors' with contents as above.

See also

Author

Tim Taylor

Examples

x = rnorm(100, mean = 0)
y = rpois(n = 100, lambda = exp(1.5 + 0.5*x))
dat <- data.frame(x = x, y = y)
poisson_model <- glm_model(y ~ x , family = "poisson")
predict(poisson_model, dat)

#> <trending_predict_tbl> 1 x 3
#>   result                          warnings errors
#>   <list>                          <list>   <list>
#> 1 <trending_prediction [100 × 7]> <NULL>   <NULL>

predict(poisson_model, dat, as_tibble = FALSE)

#> $result
#> <trending_prediction> 100 x 7
#>          x     y estimate lower_ci upper_ci lower_pi upper_pi
#>      <dbl> <int>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
#>  1 -0.780      3     3.28     2.90     3.73        0        8
#>  2 -1.78       2     2.01     1.66     2.43        0        6
#>  3 -0.428      2     3.91     3.51     4.35        0        9
#>  4 -2.03       2     1.77     1.44     2.18        0        5
#>  5  2.75      21    18.7     15.3     22.8         8       33
#>  6  1.51      14    10.2      9.02    11.5         4       19
#>  7  0.0340     2     4.91     4.48     5.38        1       10
#>  8  0.741      5     6.95     6.35     7.61        2       13
#>  9 -1.27       2     2.58     2.21     3.01        0        7
#> 10 -0.164      4     4.45     4.04     4.91        1       10
#> # … with 90 more rows
#> 
#> $warnings
#> NULL
#> 
#> $errors
#> NULL
#> 
#> attr(,"class")
#> [1] "trending_predict" "list"